requirements.txt
crewai
crewai[tools]
litellm
langchain-community
streamlit
python-dotenv

